requirements.txt
# AI Robo-Advisor Requirements

# Core libraries
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0

# Portfolio optimization
cvxpy>=1.4.0

# LLM integration
openai>=1.0.0

# Market data
yfinance>=0.2.0

# Environment variables
python-dotenv>=1.0.0
